polish: keep moe skip test; move sharding configs to avoid circular imports;
