+++ tests/unittest/_torch/auto_deploy/unit/singlegpu/transformations/library/test_moe_fusion.py
@@ -277,10 +277,13 @@ class MoEPatternModel(nn.Module):
             torch.ops.auto_deploy.torch_quant_fp4_moe,
             0.05,
             0.01,
-            marks=pytest.mark.skipif(
-                not fp4_compatible() or not trtllm_ops_available(),
-                reason="Requires FP4 + TRTLLM support",
-            ),
+            marks=[
+                pytest.mark.skipif(
+                    not fp4_compatible() or not trtllm_ops_available(),
+                    reason="Requires FP4 + TRTLLM support",
+                ),
+                pytest.mark.skip("https://nvbugs/5410946"),
+            ],
             id="fp4",
         ),
     ],
